Move all spec to spec.rs
