+++ sgl-router/src/protocols/mod.rs
@@ -1,7 +1,5 @@
 // Protocol definitions and validation for various LLM APIs
 // This module provides a structured approach to handling different API protocols
 
-pub mod common;
-pub mod generate;
-pub mod openai;
+pub mod spec;
 pub mod validation;
