--- sgl-router/src/protocols/openai/chat/mod.rs
@@ -1,13 +1,0 @@
-// Chat Completions API module
-
-pub mod request;
-pub mod response;
-pub mod types;
-pub mod validation;
-
-// Re-export main types for convenience
-pub use request::ChatCompletionRequest;
-pub use response::{
-    ChatChoice, ChatCompletionResponse, ChatCompletionStreamResponse, ChatStreamChoice,
-};
-pub use types::*;
