--- sgl-router/src/protocols/openai/mod.rs
@@ -1,8 +1,0 @@
-// OpenAI protocol module
-// This module contains all OpenAI API-compatible types and future validation logic
-
-pub mod chat;
-pub mod common;
-pub mod completions;
-pub mod errors;
-pub mod responses;
