--- sgl-router/src/protocols/openai/responses/mod.rs
@@ -1,10 +1,0 @@
-// Responses API module
-
-pub mod request;
-pub mod response;
-pub mod types;
-
-// Re-export main types for convenience
-pub use request::ResponsesRequest;
-pub use response::ResponsesResponse;
-pub use types::*;
